solve-all script

+++ vplay.js
@@ -4,29 +4,38 @@
  */
 
 const fs = require('fs');
+const md5 = require('md5');
 const fspath = require('path');
 const mkdirp = require('mkdirp');
-const clargs = require('command-line-args');
 const tsumego = require('tsumego.js');
 
-const args = clargs([
-    { name: 'input', type: String },
-    { name: 'output', type: String }, // e.g. /foo/bar/*.sgf
-]);
+const [, , inputFile, outputDir] = process.argv;
 
 try {
-    const sgf = tsumego.SGF.parse(fs.readFileSync(args.input, 'utf8'));
+    const sgf = tsumego.SGF.parse(fs.readFileSync(inputFile, 'utf8'));
     const board = new tsumego.Board(sgf);
 
+    let nvars = 0;
+    let ndups = 0;
+
     for (const move of expand(board, sgf.vars)) {
-        const path = args.output.replace('*', tsumego.hex(board.hash));
-        console.log(path);
-        mkdirp.sync(fspath.dirname(path));
-        const data = board.sgf.slice(0, -1)
-            + 'MA[' + sgf.steps[0].MA[0] + ']'
-            + 'TR' + move.slice(1) + ')';
-        fs.writeFileSync(path, data, 'utf8');
+        const hash = md5(board.sgf).slice(0, 7);
+        const path = fspath.join(outputDir, hash + '.sgf');
+
+        nvars++;
+
+        if (fs.existsSync(path)) {
+            ndups++;
+        } else {
+            mkdirp.sync(fspath.dirname(path));
+            const data = board.sgf.slice(0, -1)
+                + 'MA[' + sgf.steps[0].MA[0] + ']'
+                + 'TR' + move.slice(1) + ')';
+            fs.writeFileSync(path, data, 'utf8');
+        }
     }
+
+    console.log('vars = ' + nvars + ' dups = ' + ndups);
 } catch (err) {
     throw err;
 }
